Charlie/cleanup

+++ src/Components/SphereCollider.hpp
@@ -33,57 +33,150 @@ public:
      * @param deltaTime
      */
     void Update(float deltaTime);
+    /**
+     * @brief Updates Sphere collider
+     * @param deltaTime 
+     */
     void LateUpdate(float deltaTime);
-
+    /**
+     * @brief Updates the Phsyics on sphere collider
+     * @param deltaTime 
+     */
     void PhysicsUpdate(float deltaTime);
-
+    /**
+     * @brief Set the Orientation of the sphere
+     * @param orientation 
+     */
     void SetOrientation(glm::quat orientation);
-
+    /**
+     * @brief Get the Orientation of the sphere
+     * @return glm::quat 
+     */
     glm::quat GetOrientation();
-
+    /**
+     * @brief Set the Mass
+     * @param m 
+     */
     void SetMass(double m);
+    /**
+     * @brief Get the Mass
+     * @return double 
+     */
     double GetMass();
-
+    /**
+     * @brief Get the Inverse Mass
+     * @return double 
+     */
     double GetInverseMass();
-
-    void SetCentreOfMass(glm::dvec3 com);
-    glm::dvec3 GetCentreOfMass();
-
+    /**
+     * @brief Set the Centre Of Mass
+     * @param com 
+     */
+    void SetCentreOfMass(glm::vec3 com);
+    /**
+     * @brief Get the Centre Of Mass
+     * @return glm::dvec3 
+     */
+    glm::vec3 GetCentreOfMass();
+    /**
+     * @brief Initialize the inertia tensor
+     */
     void SetInertiaTensor();
-    glm::dmat3x3 GetInertiaTensor();
-
-    glm::dmat3x3 GetInverseInertiaTensor();
-
-    void SetLinearVelocity(glm::dvec3 lv);
-    glm::dvec3 GetLinearVelocity();
-
-    void SetAngularVelocity(glm::dvec3 av);
+    /**
+     * @brief Get the Inertia Tensor
+     * @return glm::dmat3x3 
+     */
+    glm::mat3x3 GetInertiaTensor();
+    /**
+     * @brief Get the Inverse Inertia Tensor
+     * @return glm::dmat3x3 
+     */
+    glm::mat3x3 GetInverseInertiaTensor();
+    /**
+     * @brief Set the Linear Velocity
+     * @param lv 
+     */
+    void SetLinearVelocity(glm::vec3 lv);
+    /**
+     * @brief Get the Linear Velocity
+     * @return glm::dvec3 
+     */
+    glm::vec3 GetLinearVelocity();
+    /**
+     * @brief Set the Angular Velocity
+     * @param av 
+     */
+    void SetAngularVelocity(glm::vec3 av);
+    /**
+     * @brief Get the Angular Velocity
+     * @return glm::dvec3 
+     */
     glm::dvec3 GetAngularVelocity();
-
+    /**
+     * @brief Set Is Static Object
+     * @param s 
+     */
     void SetIsStaticObject(bool s);
+    /**
+     * @brief Get Is Static Object
+     * @return bool
+     */
     bool GetIsStaticObject();
-
+    /**
+     * @brief Set is Gravity Affected
+     * @param g 
+     */
     void SetGravityAffected(bool g);
+    /**
+     * @brief Get is Gravity Affected
+     * @param g 
+     */
     bool GetGravityAffected();
-
+    /**
+     * @brief Get the Radius
+     * @return double 
+     */
     double GetRadius();
-
     /**
      * @brief Set the Position of the Sphere
-     * @param orientation
+     * @param glm::vec3
      */
-    void SetPosition(glm::dvec3 newPosition);
-
-    glm::dvec3 GetPosition();
-
+    void SetPosition(glm::vec3 newPosition);
+    /**
+     * @brief Get the Position 
+     * 
+     * @return glm::vec3 
+     */
+    glm::vec3 GetPosition();
+    /**
+     * @brief Get the Collider ID
+     * @return int 
+     */
     int GetColliderID();
-
-    void Translate(glm::dvec3 position);
-
+    /**
+     * @brief Translate the ball along a heading
+     * @param position 
+     */
+    void Translate(glm::vec3 position);
+    /**
+     * @brief Initialise static object
+     */
     void StaticSet();
-
-    void Rotate(glm::dvec3 angVelocity, float deltaTime);
+    /**
+     * @brief Rotates an object along a heading
+     * @param angVelocity 
+     * @param deltaTime 
+     */
+    void Rotate(glm::vec3 angVelocity, float deltaTime);
+    /**
+     * @brief Set the Collision Category
+     * @param category 
+     */
     void SetCollisionCategory(unsigned short category);
+    /**
+     * @brief Set the Collision Mask Bits
+     * @param maskBits 
+     */
     void SetCollisionMaskBits(unsigned short maskBits);
 
 private:
